+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## v5 (2020-01-02)
+
+- Fixed a bug so code before !DOCTYPE is handled properly
+
 ## v4 (2019-12-30)
 
 - Fixed issue where submit/send button didn't keep the intended classes